Added syncing title system, profile display card (on click), and syncing info update to card update
I get the idea! Basically, if the clients have the same function, then you can just send data over and call the respective functions with RPC (which will be called on all clients' end), so the update carries over, in a sense that data is being transmitted and processed similar across all clients' ends! So as long as data's transferred through RPC, it's free for you to decide how you'd process the data after receiving them.

Also I need to add an "add friend" button later in the friends system LOL